+++ internal/model/qr_sign_in_days.gen.go
@@ -14,10 +14,10 @@ const TableNameQrSignInDay = "qr_sign_in_days"
 
 // QrSignInDay 签到天数关联表
 type QrSignInDay struct {
-	ID         int64                 `gorm:"column:ID;primaryKey;autoIncrement:true;comment:主键" json:"id"`                        // 主键
+	ID         int32                 `gorm:"column:ID;primaryKey;autoIncrement:true;comment:主键" json:"id"`                        // 主键
 	Qq         string                `gorm:"column:QQ;not null;comment:用户QQ" json:"qq"`                                           // 用户QQ
-	MonthDays  int64                 `gorm:"column:MONTH_DAYS;not null;comment:月连续签到天数" json:"monthDays"`                         // 月连续签到天数
-	TotalDays  int64                 `gorm:"column:TOTAL_DAYS;not null;comment:总天数" json:"totalDays"`                             // 总天数
+	MonthDays  int32                 `gorm:"column:MONTH_DAYS;not null;comment:月连续签到天数" json:"monthDays"`                         // 月连续签到天数
+	TotalDays  int32                 `gorm:"column:TOTAL_DAYS;not null;comment:总天数" json:"totalDays"`                             // 总天数
 	UpdateDate *db.LocalDateTime     `gorm:"column:UPDATE_DATE;autoUpdateTime;comment:更新日期" json:"updateDate"`                    // 更新日期
 	CreateDate *db.LocalDateTime     `gorm:"column:CREATE_DATE;autoCreateTime;comment:创建日期" json:"createDate"`                    // 创建日期
 	CreateBy   *string               `gorm:"column:create_by;comment:创建人" json:"createBy"`                                        // 创建人
